GPLIM-6607, GPLIM-6681 swap double for BigDecimal.  Fixing Build breaks that come with the change.  Still need to fix test case breaks.  Also make number of lanes and quantity on customization able to be fractional.

+++ mercury/src/test/java/org/broadinstitute/gpinformatics/athena/entity/orders/ProductOrderTest.java
@@ -386,7 +386,7 @@ public class ProductOrderTest {
     public void testNonAbandonedCount() throws Exception {
         ProductOrder testParentOrder = new ProductOrder(TEST_CREATOR, "Test order with Abandoned Count",
                 sixMercurySamplesNoDupes, QUOTE, null, null);
-        testParentOrder.addSapOrderDetail(new SapOrderDetail("testParentNumber", testParentOrder.getNonAbandonedCount(),
+        testParentOrder.addSapOrderDetail(new SapOrderDetail("testParentNumber", testParentOrder.getNonAbandonedCount().intValue(),
                 testParentOrder.getQuoteId(), SapIntegrationClientImpl.SAPCompanyConfiguration.BROAD.getCompanyCode()));
 
         Assert.assertEquals(testParentOrder.getTotalNonAbandonedCount(ProductOrder.CountAggregation.ALL),6);
